requirements.txt
annotated-types==0.7.0
anyio==4.11.0
backoff==2.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
diskcache==5.6.3
distro==1.9.0
et_xmlfile==2.0.0
FalkorDB==1.2.0
graphiti-core==0.22.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.0
lxml==6.0.2
neo4j==6.0.2
numpy==2.3.3
openai==2.3.0
openpyxl==3.1.5
posthog==6.7.7
pydantic==2.12.2
pydantic-settings==2.11.0
pydantic_core==2.41.4
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
redis==5.3.1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
